# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "http://80.78.246.59/Stech/public/api/StartOnline": {
            "post": {
                "tags": [
                    "Active And No Active role_id = 3"
                ],
                "summary": "StartOnline",
                "description": "StartOnline",
                "operationId": "StartOnline",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "http://80.78.246.59/Stech/public/api/isOnlineUser": {
            "post": {
                "tags": [
                    "Active And No Active role_id = 3"
                ],
                "summary": "isOnlineUser",
                "description": "isOnlineUser",
                "operationId": "isOnlineUser",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "http://80.78.246.59/Stech/public/api/SearchCategoryName": {
            "post": {
                "tags": [
                    "Search"
                ],
                "summary": "SearchCategoryName",
                "description": "SearchCategoryName",
                "operationId": "SearchCategoryName",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "category_name": {
                                        "type": "string",
                                        "format": "category_name",
                                        "example": "Строительная техника"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "http://80.78.246.59/Stech/public/api/FilterUsers": {
            "post": {
                "tags": [
                    "Search"
                ],
                "summary": "FilterUsers",
                "description": "FilterUsers",
                "operationId": "FilterUsers",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "string",
                                        "format": "category_id",
                                        "example": "1"
                                    },
                                    "sub_category_id": {
                                        "type": "string",
                                        "format": "sub_category_id",
                                        "example": "2"
                                    },
                                    "latitude": {
                                        "type": "string",
                                        "format": "latitude",
                                        "example": "40.1817"
                                    },
                                    "longitude": {
                                        "type": "string",
                                        "format": "longitude",
                                        "example": "44.5099"
                                    },
                                    "radius": {
                                        "type": "string",
                                        "format": "radius",
                                        "example": "10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "http://80.78.246.59/Stech/public/api/OnePageRoleId3": {
            "post": {
                "tags": [
                    "Search"
                ],
                "summary": "OnePageRoleId3",
                "description": "OnePageRoleId3",
                "operationId": "OnePageRoleId3",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "string",
                                        "format": "user_id",
                                        "example": "5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    }
}